+++ bpaf_derive/src/field_tests.rs
@@ -321,6 +321,18 @@ fn many_catch() {
 }
 
 #[test]
+fn collect_catch() {
+    let input: NamedField = parse_quote! {
+        #[bpaf(argument("FILE"), collect, catch)]
+        files: Vec<std::path::PathBuf>
+    };
+    let output = quote! {
+        ::bpaf::long("files").argument::<std::path::PathBuf>("FILE").collect().catch()
+    };
+    assert_eq!(input.to_token_stream().to_string(), output.to_string());
+}
+
+#[test]
 fn option_catch() {
     let input: NamedField = parse_quote! {
         #[bpaf(argument("FILE"), optional, catch)]